# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Gateway API
  description: Agent 网关中台后端，提供账户、项目、服务管理和模型网关接口。
  version: 0.1.0
paths:
...
  /v1/chat/completions:
    post:
      tags:
      - LLM Gateway
      summary: 调用 OpenAI 兼容 Chat Completions 网关
      description: stream=false 返回完整 JSON；stream=true 在上游响应头成功后返回 SSE，并逐帧转发上游数据。下游响应开始后发生的上游错误通过
        SSE error 帧通知。
      operationId: chat_completions_v1_chat_completions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ChatCompletionRequest:
      properties:
        model:
          type: string
          maxLength: 200
          minLength: 1
          title: Model
          description: 模型标识；供应商连接原样接收此名称，必须填写供应商支持的模型名。
        messages:
          items:
            $ref: '#/components/schemas/ChatMessageRequest'
          type: array
          maxItems: 100
          minItems: 1
          title: Messages
        max_tokens:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 1.0
          - type: 'null'
          title: Max Tokens
          description: 兼容旧参数并用于额度预留。
        max_completion_tokens:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 1.0
          - type: 'null'
          title: Max Completion Tokens
          description: 生成 token 上限，也用于额度预留。
        temperature:
          anyOf:
          - type: number
            maximum: 2.0
            minimum: 0.0
          - type: 'null'
          title: Temperature
          description: 采样温度；由上游模型决定是否支持。
        top_p:
          anyOf:
          - type: number
            maximum: 1.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Top P
          description: 核采样参数；由上游模型决定是否支持。
        presence_penalty:
          anyOf:
          - type: number
            maximum: 2.0
            minimum: -2.0
          - type: 'null'
          title: Presence Penalty
          description: Presence penalty；具体支持范围以供应商为准。
        frequency_penalty:
          anyOf:
          - type: number
            maximum: 2.0
            minimum: -2.0
          - type: 'null'
          title: Frequency Penalty
          description: Frequency penalty；具体支持范围以供应商为准。
        stop:
          anyOf:
          - type: string
          - items:
              type: string
            type: array
          - type: 'null'
          title: Stop
          description: 停止生成序列；按原样转发。
        response_format:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Response Format
          description: 如 json_object 或 json_schema；按原样转发给上游。
        tools:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Tools
          description: OpenAI function tools 定义；按原样转发给上游。
        tool_choice:
          anyOf:
          - type: string
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Tool Choice
          description: 工具选择策略；按原样转发。
        parallel_tool_calls:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Parallel Tool Calls
          description: 是否允许并行工具调用。
        logit_bias:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Logit Bias
          description: token 概率偏置映射。
        logprobs:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Logprobs
          description: 是否返回 token 对数概率。
        top_logprobs:
          anyOf:
          - type: integer
            maximum: 20.0
            minimum: 0.0
          - type: 'null'
          title: Top Logprobs
          description: 返回的候选 token 概率数量。
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
          description: 尽可能确定性采样的随机种子。
        user:
          anyOf:
          - type: string
          - type: 'null'
          title: User
          description: 可选的调用方终端用户标识。
        stream:
          type: boolean
          title: Stream
          description: 是否以 Server-Sent Events 返回。
          default: false
        stream_options:
          anyOf:
          - $ref: '#/components/schemas/StreamOptionsRequest'
          - type: 'null'
          description: 流式响应选项；支持 include_usage。
        n:
          anyOf:
          - type: integer
            maximum: 8.0
            minimum: 1.0
          - type: 'null'
          title: N
          description: 生成选择数量；最多 8 个。
      additionalProperties: true
      type: object
      required:
      - model
      - messages
      title: ChatCompletionRequest
      description: 网关 Chat Completions 请求；额外标准和供应商兼容参数会透传。
    ChatMessageRequest:
      properties:
        role:
          type: string
          maxLength: 32
          minLength: 1
          title: Role
          description: 消息角色。
        content:
          title: Content
          description: 消息内容；文本、多模态数组或供应商兼容格式。
      additionalProperties: true
      type: object
      required:
      - role
      - content
      title: ChatMessageRequest
      description: 兼容 Chat Completions 的消息，允许 tool 和多模态等扩展字段。
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StreamOptionsRequest:
      properties:
        include_usage:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Include Usage
          description: 为 true 时在 SSE 结束前发送 token 用量 chunk。
      additionalProperties: false
      type: object
      title: StreamOptionsRequest
      description: 网关支持的流式选项；供应商标准中尚未实现的选项会被拒绝。
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer